Fix pgm_read_ptr in AVR pgmspace.h
Originally, `pgm_read_ptr` used to cast its argument to `const void *`, i.e. a pointer to read-only data.  
This is incorrect, because the argument is a pointer to a pointer in "flash memory". You cannot cast it to
`const void *` and then dereference it, because `void` is not an object type.
Also, the pointer itself is read-only, but the data could in theory be mutable.  
The correct type should be `void *const *`, i.e. a pointer to a read-only pointer to any data.