Properly print Arduino String objects in tests
This ensures that checks like:

    REQUIRE(some_str == "ABC");

Actually print the value of some_str instead of just `1` (which is the
String object converted to StringIfHelperType).

This is implemented by adding a specialization of the Catch StringMaker
template, so this only takes effect when StringPrinter.h is included.
This commit includes it in all the String testcases (even ones that do
not strictly use it now) for good measure.