cores: arduino: Define analogReference as weak function
Change analogReference declaration as a weak function.
This is to accommodate the module configuration in the
Arduino repository,
but it does not have any adverse effects and is a way to
coexist with the Zephyr repository, so we will include it.

c8e82c14195a74406ac3bd4c532c1bbe7a233f2e

Co-Authored-by: Martino Facchin <m.facchin@arduino.cc>
Signed-off-by: TOKITA Hiroshi <tokita.hiroshi@gmail.com>